Initialize planets and draw on canvas via seesaw
In planets.clj, implement planet constructor and planet-map constructor
as well as method to update planet population and update planet-map.

In ui.clj, implement basic UI consisting of a frame holding border-panel
with the planet-map painted on a canvas in the center. All planets are
intially represented as green circles

In key-controls.clj, implement very basic key-mapping to simulate the
passing of a year and update the planet-map model.